SPOCAN-43 cambio tipos de datos nativos por clases en dtos de entrada
