+++ src/main/java/com/spochi/dto/InitiativeResponseDTO.java
@@ -1,4 +1,4 @@
-package com.spochi.DTO;
+package com.spochi.dto;
 
 public class InitiativeResponseDTO {
